requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.3.27
langchain-community==0.3.27
langchain-mistralai==0.2.8
pydantic==2.11.7
python-dotenv==1.0.1
httpx==0.25.2
python-multipart==0.0.6
gunicorn==21.2.0
